Servo work
more work on the servo stuff, just kinda figured out how they work and what their deal is and setup #defines to handle most of the stuff. still have to allow it to set the speed of the servo and stuff like that, maybe a gradual slow down too but i dont know.